feat: created the entrylog entity
The entrylog holds all the append data to the file.
In a separated function I calculate the CRC of the content.

The content is described here:

// EntryLog represents the on-disk format of a key-value pair
//// Format according to paper:////
+-------+------------+---------+-----------+---------+-----------+ |
CRC  | Timestamp | KeySize | ValueSize |   Key   |   Value   |
+-------+------------+---------+-----------+---------+-----------+ |
 4B   |    8B     |   4B    |    4B     | KeySize | ValueSize|

For encoding all the data we use little-endian, I expect the data for
the encoding will grow from L -> R, to less to more important, as value
sitts at the rightmost. Also, is used in amd64, arm etc. Convention.